requirements.txt
flask==3.0.2
flask-cors==4.0.0
flask-socketio==5.3.6
numpy==1.26.4
pandas==2.2.1
plotly==5.19.0
python-dotenv==1.0.1
python-engineio==4.9.0
python-socketio==5.11.1
requests==2.31.0
eventlet==0.35.2
pyserial==3.5  # For serial communication with Pico W
mock==5.1.0  # For mocking MicroPython modules in development 